% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire_bonjour.R
\name{dire_bonjour}
\alias{dire_bonjour}
\title{Dire bonjour}
\usage{
dire_bonjour(prenom = "toi")
}
\arguments{
\item{prenom}{Le prénom de la personne, et par défaut "toi".}
}
\value{
salutation.
}
\description{
Dire bonjour
}
\examples{
dire_bonjour("Lisa")
dire_bonjour()
}
