% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pays.R
\docType{data}
\name{pays}
\alias{pays}
\title{Pays Dataset}
\format{
Un data frame avec 6 colonnes :
\describe{
\item{X1}{Colonne 1}
\item{X4}{Colonne 2}
\item{AF}{Code pays}
\item{AFG}{Code pays}
\item{Afghanistan}{Nom du pays en anglais}
\item{Afghanistan.1}{Nom du pays en français}
}
}
\source{
\url{lien_vers_la_source}
}
\usage{
pays
}
\description{
Un dataset contenant des informations sur les pays.
}
\keyword{datasets}
